datastream_tester/requirements.txt
functions-framework==3.3.0
SQLAlchemy==2.0.0
google-cloud-pubsub==2.14.0
pg8000==1.29.4
